+++ packages/core/src/core/const/types.ts
@@ -1,5 +1,5 @@
 import { Response } from '../../_shared/types'
 
-type ConstValue = string | number | boolean | Record<string, unknown>
+export type ConstValue = string | number | boolean | Record<string, unknown>
 
 export type ConstConfig = Record<string, ConstValue | ( () => Response<ConstValue> )>