+++ packages/plugin/todo/src/main.ts
@@ -13,15 +13,14 @@ export {
  *
  * @param   {Config}       [params] - Configuration for the plugin.
  * @returns {DovenvConfig}          - The plugin.
- *                                  import { defineConfig } from '@dovenv/core'
- *                                  import { todoPlugin } from '@dovenv/todo'
- *                                  export default defineConfig(
- *                                  todoPlugin( {
- *                                  example : {
- *                                  input : [ 'src/*.{ts,tsx,js,jsx,md}' ],
- *                                  },
- *                                  } ),
- *                                  ).
+ * @example
+ * import { defineConfig } from '@dovenv/core'
+ * import { todoPlugin } from '@dovenv/todo'
+ * export default defineConfig( todoPlugin( {
+ *     example : {
+ *       input : [ 'src/*.{ts,tsx,js,jsx,md}' ],
+ *     },
+ *  } )).
  */
 export const todoPlugin = ( params?: Config ): DovenvConfig => {