+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/core
 
+## 1.5.4
+
+### Patch Changes
+
+- up deps, increase perf
+
+- Updated dependencies []:
+  - @dovenv/utils@1.5.4
+
 ## 1.5.3
 
 ### Patch Changes