+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dovenv/utils
 
+## 1.5.4
+
+### Patch Changes
+
+- up deps, increase perf
+
 ## 1.5.3
 
 ### Patch Changes